KeepAlive karabiner_grabber (from LaunchAgents)

+++ src/core/grabber/src/main.cpp
@@ -19,8 +19,57 @@ int main(int argc, const char* argv[]) {
   signal(SIGUSR1, SIG_IGN);
   signal(SIGUSR2, SIG_IGN);
 
+  //
+  // Check euid
+  // (grabber is launched from LaunchDaemons (root) and LaunchAgents (user).)
+  //
+
+  bool root = (geteuid() == 0);
+
+  //
+  // Setup logger
+  //
+
+  if (root) {
+    krbn::logger::set_async_rotating_logger("grabber",
+                                            "/var/log/karabiner/grabber.log",
+                                            0755);
+  } else {
+    krbn::logger::set_async_rotating_logger("grabber_agent",
+                                            krbn::constants::get_user_log_directory() + "/grabber_agent.log",
+                                            0700);
+  }
+
+  krbn::logger::get_logger()->info("version {0}", karabiner_version);
+
+  //
+  // Check another process
+  //
+
+  {
+    bool another_process_running = false;
+
+    if (root) {
+      std::string pid_file_path = krbn::constants::get_pid_directory() + "/karabiner_grabber.pid";
+      another_process_running = !krbn::process_utility::lock_single_application(pid_file_path);
+    } else {
+      another_process_running = !krbn::process_utility::lock_single_application_with_user_pid_file("karabiner_grabber_agent.pid");
+    }
+
+    if (another_process_running) {
+      auto message = "Exit since another process is running.";
+      krbn::logger::get_logger()->info(message);
+      std::cerr << message << std::endl;
+      return 1;
+    }
+  }
+
+  //
+  // Prepare state_json_writer
+  //
+
   std::shared_ptr<krbn::state_json_writer> state_json_writer;
-  if (geteuid() == 0) {
+  if (root) {
     state_json_writer = std::make_shared<krbn::state_json_writer>(
         krbn::constants::get_grabber_state_json_file_path());
   }
@@ -34,49 +83,21 @@ int main(int argc, const char* argv[]) {
     if (state_json_writer) {
       state_json_writer->set("hid_device_open_permitted", false);
     }
-
-    return 0;
-  }
-
-  if (state_json_writer) {
-    state_json_writer->set("hid_device_open_permitted", true);
-  }
-
-  //
-  // Check euid
-  //
-
-  if (geteuid() != 0) {
-    return 0;
-  }
-
-  // Setup logger
-
-  krbn::logger::set_async_rotating_logger("grabber",
-                                          "/var/log/karabiner/grabber.log",
-                                          0755);
-
-  krbn::logger::get_logger()->info("version {0}", karabiner_version);
-
-  // Check another process
-
-  {
-    std::string pid_file_path = krbn::constants::get_pid_directory() + "/karabiner_grabber.pid";
-    if (!krbn::process_utility::lock_single_application(pid_file_path)) {
-      std::string message("Exit since another process is running.");
-      krbn::logger::get_logger()->info(message);
-      std::cerr << message << std::endl;
-      return 1;
+  } else {
+    if (state_json_writer) {
+      state_json_writer->set("hid_device_open_permitted", true);
     }
   }
 
+  //
   // Make socket directory.
+  //
 
-  krbn::filesystem_utility::mkdir_tmp_directory();
+  if (root) {
+    krbn::filesystem_utility::mkdir_tmp_directory();
 
-  unlink(krbn::constants::get_grabber_socket_file_path());
+    unlink(krbn::constants::get_grabber_socket_file_path());
 
-  {
     // Make console_user_server_socket_directory
 
     std::stringstream ss;
@@ -88,23 +109,32 @@ int main(int argc, const char* argv[]) {
     chmod(krbn::constants::get_console_user_server_socket_directory(), 0755);
   }
 
+  //
   // Run components_manager
+  //
 
-  std::shared_ptr<krbn::grabber::components_manager> components_manager;
+  // We have to use raw pointer (not smart pointer) to delete it in `dispatch_async`.
+  krbn::grabber::components_manager* components_manager = nullptr;
 
   auto version_monitor = std::make_shared<krbn::version_monitor>(krbn::constants::get_version_file_path());
 
-  version_monitor->changed.connect([&](auto&& version) {
+  version_monitor->changed.connect([&components_manager](auto&& version) {
     dispatch_async(dispatch_get_main_queue(), ^{
-      components_manager = nullptr;
+      if (components_manager) {
+        delete components_manager;
+      }
       CFRunLoopStop(CFRunLoopGetCurrent());
     });
   });
 
-  components_manager = std::make_shared<krbn::grabber::components_manager>(version_monitor);
+  if (root) {
+    components_manager = new krbn::grabber::components_manager(version_monitor);
+  }
 
   version_monitor->async_start();
-  components_manager->async_start();
+  if (components_manager) {
+    components_manager->async_start();
+  }
 
   CFRunLoopRun();
